clean up shared types and template deps
