+++ src/tasks/prisma-postgres.ts
@@ -1,6 +1,6 @@
 import { execa } from "execa";
 
-import type { PackageManager, PrismaPostgresResult } from "../types";
+import type { PackageManager } from "../types";
 import { getPackageExecutionArgs, getPackageExecutionCommand } from "../utils/package-manager";
 
 type CreateDbJsonPayload = {
@@ -11,6 +11,11 @@ type CreateDbJsonPayload = {
   databaseUrl?: string;
   claimUrl?: string;
   claimURL?: string;
+};
+
+type PrismaPostgresResult = {
+  databaseUrl: string;
+  claimUrl?: string;
 };
 
 export const PRISMA_POSTGRES_TEMPORARY_NOTICE =
