+++ templates/create/turborepo/apps/api/package.json.hbs
@@ -10,12 +10,10 @@
   },
   "dependencies": {
     "@hono/node-server": "^1.19.9",
-    "@repo/db": "{{#if (eq packageManager "npm")}}*{{else}}workspace:*{{/if}}",
     "hono": "^4.12.2"
   },
   "devDependencies": {
     "@types/node": "^24.3.0",
-    "tsx": "^4.7.1",
     "typescript": "^5.9.3"
   }
 }
